Chain comparators correctly
